Bug fix - displaying ROIs in View windows
I wasn't extracting the information stored in the tuple correctly
This stored the view windows that had already been assigned to the SegmentationDisplayNode.
I wanted to append any previous assignments to a new assignment.
Eg. If T1W is displayed already on Red node, and I also want it displayed on the Green node, I have to reassign using a list containing both nodes.