Remove DicomWidgets override
No longer needed with the newer Slicer version. This was added to the project to override the pop-up asking if you want all associated Dicom  series to be loaded with the selected image volume.
This may be handled in the SlicerRT extension which is required to be downloaded for the Image Quizzer
Also added some general cleanup.