Warn if no label maps created
If segmentation is required, the user has the option to not segment if there are no lesions. The system will warn if there are no label map nodes detected.
The warning has OK/Cancel to allow user to go back to segment if the Next button was pressed by accident.
This warning is initiated through the PerformSave function. The OK/Cancel warning is bypassed if PerformSave is called from the 'X' exit button. (There is no cancel).