RotateToVolumePlane with 'Acquisition' orientation
If xml defines the image orientation to be 'Acquisition', the volume is rotated to the plane that the image was acquired. This allows for 'clean' contouring of ROI's. 
However, before rotating, the code has to 'anchor' the image closest to the plane in which the image was acquired (Axian, Sagittal, Coronal). This  is computed from the ijkToRAS matrix stored with the volume. The viewing widget is set to that orientation. This is like an anchor to the closest plane. Then if the patient was slightly rotated from the true plane, the RotationToVolume adjusts the image view.