Copy previous label map to current image
If usepreviouslabelmap attribute is set to 'y' for an image, the code will search through history for an identical image and copy it's label map (created by the user) if there is one stored to the current image.
The user can then edit the label map if desired.
A previous label map is not copied if one already exists for the current image.