Bug Fix for RotateToVolume - turn off link control
Wait until all images have been assigned with orientation, destination and label maps (in AssignNodesToView) before turning on the Link Control. (If link control is on during the entire process, images that were already assigned to their widgets with the requested orientation may be reset by subsequent image assignments.)
Add question set 'id' to the group box title (an extra descriptor).
Add page name and descriptor to the progress bar. It was put in the progress bar to save on 'real estate' in the quiz tab widget.