Add ROI Color table attribute
At the Session (root) level define the ROI color table that is to be used for segmenting. 
This way the admin does not have to 'remember' to adjust the QuizzerColorTable before each study is run. 
And, this provides for flexibility of having the user launch different quizzes that require different ROI color definitions.