Fix Resume if quiz exits prematurely; fix 'Finish' at en of quiz
Adding the login timestamp only after the initial write of captured responses to the xml allows for resuming to function properly. If user exits quiz prematurely, the app searches for the last question set with the latest login timestamp to determine where to resume.
Saving the responses when pressing the 'Finish' button (Next button that changes to finish at the very last question), has been fixed to check that all questions were answered and a label map created if requested.