Modify attribute names and move Image Destination to element
QuestionSet attributes changed from id and title to name and descriptor to be consistent with Page and Image attributes. (Less things to remember).
Move Image attribute 'destination' to be an element. This change is more in line with the other elements of the image : Layer and Orientation.
Delete unused development xml files.