Added Progress window for data load; and ChangeLabelEffect
For long wait times when loading multiple images (Dicom series can have numerous images), I added a progress bar window. It moves in 'chunks' because of when the processing can be interrupted to update the bar (1%, 14%, 50%, done)- but at least it is better than wondering if anything is happening.
I also opened up the segmenting tools to add the Change Label Effect - allowing the user to adjust the assigned label number from one number to another.