Add testing xml scenario
Test the 'UsePreviousLabelMap and AllowMultipleResponse attributes.
Using the next and previous buttons, these attributes are tested.
Currently the contours are copied from page 1 to page 2 ONLY ONCE - the first time the user transitions from page 1 to 2.
- If AllowMultipleResponse is set to Y, and the user changed the contours on page 1, those changes will not be copied to page 2. 
(This may change in the future to allow the program to always copy the label map from page 1 to 2 but it poses different problems going forward.)